Bugfix: Cut&paste error from the NetBSD code.
Also: Change the initialisation of the command string to a static
initialiser. Verify it against the output of umass.c when being sent a
command using 'camcontrol eject da0' to a Bulk-Only device.

This should make those devices work that need a SCSI eject command to
switch to modem mode (Novatel 950D and others).


git-svn-id: svn://svn.freebsd.org/base/head@184605 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f